# Quick patch latency issues