--- a/crates/bifrost/src/loglets/local_loglet/metric_definitions.rs
+++ b/crates/bifrost/src/loglets/local_loglet/metric_definitions.rs
@@ -1,0 +1,31 @@
+// Copyright (c) 2024 -  Restate Software, Inc., Restate GmbH.
+// All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+/// Optional to have but adds description/help message to the metrics emitted to
+/// the metrics' sink.
+use metrics::{describe_counter, describe_histogram, Unit};
+
+pub(crate) const BIFROST_LOCAL_APPEND: &str = "restate.bifrost.localloglet.appends.total";
+
+pub(crate) const BIFROST_LOCAL_APPEND_DURATION: &str =
+    "restate.bifrost.localloglet.append_duration.seconds";
+
+pub(crate) fn describe_metrics() {
+    describe_counter!(
+        BIFROST_LOCAL_APPEND,
+        Unit::Count,
+        "Number of append requests to bifrost's local loglet"
+    );
+    describe_histogram!(
+        BIFROST_LOCAL_APPEND_DURATION,
+        Unit::Seconds,
+        "Total latency of bifrost's local loglet appends"
+    );
+}