Quick patch latency issues

According to my test, this takes a simple wrk load test result from ~10rps to ~700qps

before:
```
~/w/r/sdk-java ❯❯❯ wrk -t 15 -c 25 http://localhost:8080/Counter/my-counter/get

Running 10s test @ http://localhost:8080/Counter/my-counter/get
  15 threads and 25 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.20s   230.15ms   1.34s    90.68%
    Req/Sec     0.14      0.56     4.00     91.53%
  118 requests in 10.09s, 26.16KB read
Requests/sec:     11.69
Transfer/sec:      2.59KB
```

after:
```
~/w/r/sdk-java ❯❯❯ wrk -t 15 -c 25 http://localhost:8080/Counter/my-counter/get

Running 10s test @ http://localhost:8080/Counter/my-counter/get
  15 threads and 25 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    18.01ms    2.98ms  52.11ms   62.40%
    Req/Sec    55.54      9.63    80.00     68.93%
  8370 requests in 10.07s, 1.81MB read
Requests/sec:    831.27
Transfer/sec:    184.27KB
```

---
[//]: # (BEGIN SAPLING FOOTER)
Stack created with [Sapling](https://sapling-scm.com). Best reviewed with [ReviewStack](https://reviewstack.dev/restatedev/restate/pull/1396).
* __->__ #1396
* #1395